[OWL-865][transfer] `TotalTime` should be float32
The type of connection time's value from `tcpconn` is float.

+++ modules/transfer/sender/queue_item.go
@@ -69,7 +69,7 @@ type nqmConnItem struct {
 	Timestamp int64       `json:"time"`
 	Agent     nqmEndpoint `json:"agent"`
 	Target    nqmEndpoint `json:"target"`
-	TotalTime int32       `json:"total_time"`
+	TotalTime float32     `json:"total_time"`
 }
 
 func (this nqmConnItem) String() string {
